+++ backend/src/managers/server/server-manager.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ImageDBModule } from '../../collections/image-db/image-db.module.js';
-import { ObjectStorageModule } from '../../collections/object-storage/object-storage.module.js';
+import { ExternalStorageModule } from '../../collections/external-storage/external-storage.module.js';
 import { ServerSettingsDbModule } from '../../collections/server-settings-db/server-settings-db.module.js';
 import { SystemStateDbModule } from '../../collections/system-state-db/system-state-db.module.js';
 import { EarlyConfigModule } from '../../config/early/early-config.module.js';
@@ -13,7 +13,7 @@ import { StorageMigrationService } from './storage-migration.service.js';
     ServerSettingsDbModule,
     SystemStateDbModule,
     ImageDBModule,
-    ObjectStorageModule,
+    ExternalStorageModule,
   ],
   providers: [ServerSettingsService, StorageMigrationService],
   exports: [ServerSettingsService, StorageMigrationService],